requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
asgiref==3.5.0
    # via
    #   -r requirements.in
    #   django
asttokens==2.0.5
    # via
    #   -r requirements.in
    #   stack-data
attrs==21.4.0
    # via
    #   -r requirements.in
    #   jsonschema
backcall==0.2.0
    # via
    #   -r requirements.in
    #   ipython
backports-zoneinfo==0.2.1
    # via
    #   -r requirements.in
    #   django
certifi==2021.10.8
    # via
    #   -r requirements.in
    #   requests
charset-normalizer==2.0.12
    # via
    #   -r requirements.in
    #   requests
click==8.1.2
    # via
    #   -r requirements.in
    #   pip-tools
coreapi==2.3.3
    # via
    #   -r requirements.in
    #   openapi-codec
coreschema==0.0.4
    # via
    #   -r requirements.in
    #   coreapi
decorator==5.1.1
    # via
    #   -r requirements.in
    #   ipdb
    #   ipython
django==4.0.4
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.13.1
    # via -r requirements.in
executing==0.8.3
    # via
    #   -r requirements.in
    #   stack-data
flake8==4.0.1
    # via -r requirements.in
idna==3.3
    # via
    #   -r requirements.in
    #   requests
importlib-resources==5.7.1
    # via
    #   -r requirements.in
    #   jsonschema
inflection==0.5.1
    # via -r requirements.in
ipdb==0.13.9
    # via -r requirements.in
ipython==8.2.0
    # via
    #   -r requirements.in
    #   ipdb
isort==5.10.1
    # via -r requirements.in
itypes==1.2.0
    # via
    #   -r requirements.in
    #   coreapi
jedi==0.18.1
    # via
    #   -r requirements.in
    #   ipython
jinja2==3.1.1
    # via
    #   -r requirements.in
    #   coreschema
jsonschema==4.4.0
    # via
    #   -r requirements.in
    #   swagger-spec-validator
markupsafe==2.1.1
    # via
    #   -r requirements.in
    #   jinja2
matplotlib-inline==0.1.3
    # via
    #   -r requirements.in
    #   ipython
mccabe==0.6.1
    # via
    #   -r requirements.in
    #   flake8
openapi-codec==1.3.2
    # via -r requirements.in
packaging==21.3
    # via -r requirements.in
parso==0.8.3
    # via
    #   -r requirements.in
    #   jedi
pep517==0.12.0
    # via
    #   -r requirements.in
    #   pip-tools
pexpect==4.8.0
    # via
    #   -r requirements.in
    #   ipython
pickleshare==0.7.5
    # via
    #   -r requirements.in
    #   ipython
pip-tools==6.6.0
    # via -r requirements.in
prompt-toolkit==3.0.29
    # via
    #   -r requirements.in
    #   ipython
ptyprocess==0.7.0
    # via
    #   -r requirements.in
    #   pexpect
pure-eval==0.2.2
    # via
    #   -r requirements.in
    #   stack-data
pycodestyle==2.8.0
    # via
    #   -r requirements.in
    #   flake8
pycrypto==2.6
    # via -r requirements.in
pyflakes==2.4.0
    # via
    #   -r requirements.in
    #   flake8
pygments==2.11.2
    # via
    #   -r requirements.in
    #   ipython
pyparsing==3.0.8
    # via
    #   -r requirements.in
    #   packaging
pyrsistent==0.18.1
    # via
    #   -r requirements.in
    #   jsonschema
pytz==2022.1
    # via
    #   -r requirements.in
    #   djangorestframework
pyyaml==6.0
    # via
    #   -r requirements.in
    #   swagger-spec-validator
requests==2.27.1
    # via
    #   -r requirements.in
    #   coreapi
ruamel-yaml==0.17.21
    # via -r requirements.in
ruamel-yaml-clib==0.2.6
    # via
    #   -r requirements.in
    #   ruamel-yaml
simplejson==3.17.6
    # via -r requirements.in
six==1.16.0
    # via
    #   -r requirements.in
    #   asttokens
    #   swagger-spec-validator
sqlparse==0.4.2
    # via
    #   -r requirements.in
    #   django
stack-data==0.2.0
    # via
    #   -r requirements.in
    #   ipython
swagger-spec-validator==2.7.4
    # via -r requirements.in
toml==0.10.2
    # via
    #   -r requirements.in
    #   ipdb
tomli==2.0.1
    # via
    #   -r requirements.in
    #   pep517
traitlets==5.1.1
    # via
    #   -r requirements.in
    #   ipython
    #   matplotlib-inline
uritemplate==4.1.1
    # via
    #   -r requirements.in
    #   coreapi
urllib3==1.26.9
    # via
    #   -r requirements.in
    #   requests
wcwidth==0.2.5
    # via
    #   -r requirements.in
    #   prompt-toolkit
wheel==0.37.1
    # via pip-tools
zipp==3.8.0
    # via
    #   -r requirements.in
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
